+++ qadence_protocols/measurements/utils_shadow/unitaries.py
@@ -1,0 +1,23 @@
+from __future__ import annotations
+import torch
+from qadence.operations import H, I, SDagger, X, Y, Z
+from qadence.blocks.block_to_tensor import HMAT, IMAT, SDAGMAT
+
+
+pauli_gates = [X, Y, Z]
+pauli_rotations = [
+    lambda index: H(index),
+    lambda index: SDagger(index) * H(index),
+    lambda index: None,
+]
+
+UNITARY_TENSOR = [
+    HMAT.squeeze(dim=0),
+    (HMAT @ SDAGMAT).squeeze(dim=0),
+    IMAT.squeeze(dim=0),
+]
+UNITARY_TENSOR_adjoint = [unit.adjoint() for unit in UNITARY_TENSOR]
+
+idmat = UNITARY_TENSOR[-1]
+
+HammingMatrix = torch.tensor([[1.0, -0.5], [-0.5, 1.0]], dtype=torch.double)
